requirements.txt
streamlit==1.29.0
wikipedia==1.4.0
google-generativeai==0.3.1
pillow==10.1.0
python-dotenv==1.0.0
plotly==5.18.0
pycountry==23.12.11
requests==2.31.0
